refactor: move SchemaConfig from topology crate to config crate
