+++ server/src/config.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 
 use analytic_engine;
 use ceresdbproto_deps::ceresdbproto::storage;
-use cluster::{config::ClusterConfig, topology::SchemaConfig};
+use cluster::config::{ClusterConfig, SchemaConfig};
 use common_types::schema::TIMESTAMP_COLUMN;
 use meta_client::types::ShardId;
 use serde_derive::Deserialize;
